test: also confirm after
