+++ test/cli/fixtures/custom-runner/test-runner.ts
@@ -1,21 +1,13 @@
 import type { Suite, TestContext } from '@vitest/runner'
 import { VitestTestRunner } from 'vitest/runners'
+import { getSuiteNames } from './utils';
 
 class CustomTestRunner extends VitestTestRunner {
 	extendTaskContext(context: TestContext) {
     super.extendTaskContext(context);
-		(context as any).__suiteNames = getSuiteNames(context.task.suite);
+		(context as any).__suiteNames = getSuiteNames(context.task.suite)
     return context
 	}
-}
-
-function getSuiteNames(suite?: Suite) {
-	const names = []
-	while (suite) {
-		names.push(suite.name)
-		suite = suite.suite
-	}
-	return names
 }
 
 export default CustomTestRunner
