+++ test/cli/fixtures/custom-runner/utils.ts
@@ -1,0 +1,10 @@
+import { Suite } from "@vitest/runner"
+
+export function getSuiteNames(suite?: Suite) {
+  const names = []
+  while (suite) {
+    names.push(suite.name)
+    suite = suite.suite
+  }
+  return names
+}
